# Conflicts:
#	jobdori-api/src/main/kotlin/com/jobdori/api/application/experience/service/ExperienceService.kt
#	jobdori-infrastructure/persistence/src/main/resources/data.sql

+++ jobdori-api/src/main/kotlin/com/jobdori/api/application/experience/dto/response/ExperienceProjectResponse.kt
@@ -9,11 +9,11 @@ data class ExperienceProjectResponse(
     val summary: String,
     val period: PeriodResponse?,
     val role: String?,
-    val experienceCount: Int? = null,
+    val experienceCount: Long? = null,
 ) {
 
     companion object {
-        fun from(project: ExperienceProject, experienceCount: Int? = null) = ExperienceProjectResponse(
+        fun from(project: ExperienceProject, experienceCount: Long? = null) = ExperienceProjectResponse(
             projectId = project.id,
             name = project.name,
             summary = project.summary,
